Attempt to handle tricky situation where some instances of an online account contain the nonce and recent block signature, whereas other instances do not (due to being sent via an older peer).
Right now, two OnlineAccountData objects are considered equal if they have matching timestamps, signatures, and public keys. This reduces the chance of multiple versions of the same online account data from being sent around the network. The downside is that an instance containing a nonce value can be ignored due to already having an inferior OnlineAccountData instance in the list.

The current approach is this:
- Only allow new duplicate onlineAccountData to be added to the import queue if it's superior to the one we already have.
- Remove the existing, inferior data at the time of import (once the new data is considered valid).

This is only a temporary problem, and can be simplified once the additional fields in OnlineAccountsV3Message become required rather than optional.